+++ src/firebase-app.ts
@@ -31,6 +31,8 @@ import {FirestoreService} from './firestore/firestore';
 import {InstanceId} from './instance-id/instance-id';
 import {ProjectManagement} from './project-management/project-management';
 
+import {Agent} from 'http';
+
 /**
  * Type representing a callback which is called every time an app lifecycle event occurs.
  */
@@ -46,6 +48,7 @@ export interface FirebaseAppOptions {
   serviceAccountId?: string;
   storageBucket?: string;
   projectId?: string;
+  httpAgent?: Agent;
 }
 
 /**